Разобрался со сплэшскринами. В зависимости от версии андроид запскается либо кастомный, либо дефолтный для 12 версии андроид сплэш
