+++ app/build.gradle
@@ -41,6 +41,9 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.4.2'
     implementation 'com.google.android.material:material:1.6.1'
 
+    // splashscreen
+    implementation 'androidx.core:core-splashscreen:1.0.0-rc01'
+
     // widget
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
