+++ app/src/main/AndroidManifest.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     package="com.example.newsgb">
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -14,18 +15,19 @@
         android:theme="@style/Theme.NewsGB">
         <activity
             android:name="._core.ui.MainActivity"
-            android:exported="false"
-            android:screenOrientation="portrait"/>
-        <activity
-            android:name="._core.ui.SplashScreenActivity"
             android:exported="true"
-            android:screenOrientation="portrait">
+            android:screenOrientation="portrait"
+            tools:ignore="LockedOrientationActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name="._core.ui.CustomSplashScreenActivity"
+            android:screenOrientation="portrait"
+            tools:ignore="LockedOrientationActivity" />
     </application>
 
 </manifest>
