+++ app/src/main/res/layout/activity_custom_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/splash_background"
-    tools:context="._core.ui.SplashScreenActivity">
+    tools:context="._core.ui.CustomSplashScreenActivity">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/splash_image_view"
