+++ app/src/main/res/values-night/colors.xml
@@ -4,4 +4,5 @@
     <color name="item_card_background_color">#232323</color>
     <color name="category_button_text_color">#BCACAC</color>
     <color name="separator_color">#BCACAC</color>
+    <color name="splash_background">#4C4C4C</color>
 </resources>
