+++ app/src/main/res/values/strings.xml
@@ -23,4 +23,5 @@
     <string name="dialog_title_device_is_offline">No internet connection</string>
     <string name="dialog_button_cancel">Cancel</string>
     <string name="dialog_title_stub">Undefined error</string>
+    <string name="loading">LOADING…</string>
 </resources>
